SecMan.UnitTests/Controller/AuthControllerTests.cs
using AutoFixture;
using MediatR;
using Microsoft.AspNetCore.Authentication;
using Microsoft.AspNetCore.Authentication.Cookies;
using Microsoft.AspNetCore.Http;
using Microsoft.AspNetCore.Mvc;
using Microsoft.Data.Sqlite;
using Microsoft.Extensions.Configuration;
using Moq;
using SecMan.Interfaces.BL;
using SecMan.Model;
using SecMan.UnitTests.Logger;
using Serilog;
using System.Net;
using System.Security.Claims;
using UserAccessManagement.Controllers;

namespace SecMan.UnitTests.Controller
{
    [CustomLogging]
    [Collection("Sequential Collection")]
    public class AuthControllerTests
    {
        private readonly IFixture _fixture;
        private readonly Mock<IAuthBL> _mockAuthBl;
        private readonly Mock<IPasswordBl> _mockPasswordBl;
        private readonly Mock<IConfiguration> _mockConfiguration;
        private readonly Mock<IHttpContextAccessor> _mockHttpContextAccessor;
        private readonly Mock<HttpContext> _mockHttpContext;
        private readonly AuthController _authController;
        private readonly DefaultHttpContext _httpContext;
        private readonly Mock<IMediator> _mockMediator;
        private readonly Mock<IAuthenticationService> _mockAuthenticationService;
        private readonly string _sessionId;

        public AuthControllerTests()
        {
            _fixture = new Fixture();
            _mockAuthBl = new Mock<IAuthBL>();
            _mockPasswordBl = new Mock<IPasswordBl>();
            _mockConfiguration = new Mock<IConfiguration>();
            _mockHttpContextAccessor = new Mock<IHttpContextAccessor>();
            _mockMediator = new Mock<IMediator>();
            _authController = new AuthController(_mockPasswordBl.Object, _mockConfiguration.Object, _mockHttpContextAccessor.Object, _mockAuthBl.Object, _mockMediator.Object);
            _httpContext = new DefaultHttpContext();
            _mockHttpContext = new Mock<HttpContext>();
            _mockHttpContextAccessor.Setup(x => x.HttpContext)
                .Returns(_mockHttpContext.Object);
            _mockAuthenticationService = new Mock<IAuthenticationService>();
            _mockHttpContext.Setup(x => x.RequestServices.GetService(typeof(IAuthenticationService)))
                           .Returns(_mockAuthenticationService.Object);
            _mockAuthenticationService.Setup(x => x.SignInAsync(
                            It.IsAny<HttpContext>(),
                            CookieAuthenticationDefaults.AuthenticationScheme,
                            It.IsAny<ClaimsPrincipal>(),
                            It.IsAny<AuthenticationProperties>()))
                            .Returns(Task.CompletedTask);
            _sessionId = _fixture.Create<string>();
            List<Claim> claims = new List<Claim>
                {
                    new Claim(ResponseHeaders.SSOSessionId, _sessionId),
                    new Claim(ClaimTypes.Expiration, DateTime.UtcNow.AddMinutes(30).ToString("o"))
                };
            ClaimsIdentity identity = new ClaimsIdentity(claims, CookieAuthenticationDefaults.AuthenticationScheme);
            ClaimsPrincipal claimsPrincipal = new ClaimsPrincipal(identity);
            _mockHttpContext.SetupGet(x => x.User).Returns(claimsPrincipal);
            _authController.ControllerContext = new ControllerContext
            {
                HttpContext = _mockHttpContext.Object,
            };
        }


        // LoginAsync

        [Fact]
        public async Task LoginAsync_ShouldReturn_OkResponse_WithToken()
        {
            // Arrange
            LoginRequest loginModel = _fixture.Create<LoginRequest>();
            LoginServiceResponse loginResponse = _fixture.Create<LoginServiceResponse>();
            _mockAuthBl.Setup(x => x.LoginAsync(It.IsAny<LoginRequest>()))
                .ReturnsAsync(new ServiceResponse<LoginServiceResponse>(ResponseConstants.Success, HttpStatusCode.OK, loginResponse));
            Log.Information("Completed Moqing dependencies");


            // Act
            IActionResult result = await _authController.LoginAsync(loginModel);
            Log.Information("Test result: {@Result}", result);

            // Assert
            OkObjectResult okResult = Assert.IsType<OkObjectResult>(result);
            Log.Information("Verified if the result is of OkResponse");
            LoginResponse loginResponseReturned = Assert.IsType<LoginResponse>(okResult.Value);
            Log.Information("Verified it the result body is of LoginResponse Type");
            Assert.Equal(loginResponseReturned.Token, loginResponse.Token);
            Log.Information("Verified if the mocked token is returned from action method invocation");
        }


        [Fact]
        public async Task LoginAsync_ShouldReturn_UnUnauthorizedResponse_WhenUserDoesNotExists()
        {
            // Arrange
            LoginRequest loginModel = _fixture.Create<LoginRequest>();
            _mockAuthBl.Setup(x => x.LoginAsync(It.IsAny<LoginRequest>()))
                .ReturnsAsync(new ServiceResponse<LoginServiceResponse>(ResponseConstants.UserDoesNotExists, HttpStatusCode.Unauthorized));
            Log.Information("Completed Moqing dependencies");


            // Act
            IActionResult result = await _authController.LoginAsync(loginModel);
            Log.Information("Test result: {@Result}", result);


            // Assert
            UnauthorizedObjectResult unauthorizedResult = Assert.IsType<UnauthorizedObjectResult>(result);
            Log.Information("Verified if the response is of type UnauthorizedObjectResult");
            Unauthorized loginResponseReturned = Assert.IsType<Unauthorized>(unauthorizedResult.Value);
            Log.Information("Verified if the response object is of type Unauthorized");
            Assert.Equal(ResponseConstants.UserDoesNotExists, loginResponseReturned.Detail);
            Log.Information("Verified if the response object's Detail is UserDoesNotExists");
        }



        [Fact]
        public async Task LoginAsync_ShouldReturn_UnUnauthorizedResponse_WhenUserLocked()
        {
            // Arrange
            LoginRequest loginModel = _fixture.Create<LoginRequest>();
            _mockAuthBl.Setup(x => x.LoginAsync(It.IsAny<LoginRequest>()))
                .ReturnsAsync(new ServiceResponse<LoginServiceResponse>(ResponseConstants.AccountLocked, HttpStatusCode.Unauthorized));
            Log.Information("Completed Moqing dependencies");


            // Act
            IActionResult result = await _authController.LoginAsync(loginModel);
            Log.Information("Test result: {@Result}", result);


            // Assert
            UnauthorizedObjectResult unauthorizedResult = Assert.IsType<UnauthorizedObjectResult>(result);
            Log.Information("Verified if the response is of type UnauthorizedObjectResult");
            Unauthorized loginResponseReturned = Assert.IsType<Unauthorized>(unauthorizedResult.Value);
            Log.Information("Verified if the response object is of type Unauthorized");
            Assert.Equal(ResponseConstants.AccountLocked, loginResponseReturned.Detail);
            Log.Information("Verified if the response object's Detail is AccountLocked");
        }

        [Fact]
        public async Task LoginAsync_ShouldReturn_UnUnauthorizedResponse_WhenUserRetired()
        {
            // Arrange
            LoginRequest loginModel = _fixture.Create<LoginRequest>();
            _mockAuthBl.Setup(x => x.LoginAsync(It.IsAny<LoginRequest>()))
                .ReturnsAsync(new ServiceResponse<LoginServiceResponse>(ResponseConstants.AccountRetired, HttpStatusCode.Unauthorized));
            Log.Information("Completed Moqing dependencies");


            // Act
            IActionResult result = await _authController.LoginAsync(loginModel);
            Log.Information("Test result: {@Result}", result);


            // Assert
            UnauthorizedObjectResult unauthorizedResult = Assert.IsType<UnauthorizedObjectResult>(result);
            Log.Information("Verified if the response is of type UnauthorizedObjectResult");
            Unauthorized loginResponseReturned = Assert.IsType<Unauthorized>(unauthorizedResult.Value);
            Log.Information("Verified if the response object is of type Unauthorized");
            Assert.Equal(ResponseConstants.AccountRetired, loginResponseReturned.Detail);
            Log.Information("Verified if the response object's Detail is AccountRetired");
        }


        [Fact]
        public async Task LoginAsync_ShouldReturn_UnUnauthorizedResponse_WhenUserInactive()
        {
            // Arrange
            LoginRequest loginModel = _fixture.Create<LoginRequest>();
            _mockAuthBl.Setup(x => x.LoginAsync(It.IsAny<LoginRequest>()))
                .ReturnsAsync(new ServiceResponse<LoginServiceResponse>(ResponseConstants.AccountInActive, HttpStatusCode.Unauthorized));
            Log.Information("Completed Moqing dependencies");


            // Act
            IActionResult result = await _authController.LoginAsync(loginModel);
            Log.Information("Test result: {@Result}", result);


            // Assert
            UnauthorizedObjectResult unauthorizedResult = Assert.IsType<UnauthorizedObjectResult>(result);
            Log.Information("Verified if the response is of type UnauthorizedObjectResult");
            Unauthorized loginResponseReturned = Assert.IsType<Unauthorized>(unauthorizedResult.Value);
            Log.Information("Verified if the response object is of type Unauthorized");
            Assert.Equal(ResponseConstants.AccountInActive, loginResponseReturned.Detail);
            Log.Information("Verified if the response object's Detail is AccountInActive");
        }


        [Fact]
        public async Task LoginAsync_ShouldReturn_UnUnauthorizedResponse_WhenPasswordIsExpired()
        {
            // Arrange
            LoginRequest loginModel = _fixture.Create<LoginRequest>();
            _mockAuthBl.Setup(x => x.LoginAsync(It.IsAny<LoginRequest>()))
                .ReturnsAsync(new ServiceResponse<LoginServiceResponse>(ResponseConstants.PasswordExpired, HttpStatusCode.Unauthorized));
            Log.Information("Completed Moqing dependencies");


            // Act
            IActionResult result = await _authController.LoginAsync(loginModel);
            Log.Information("Test result: {@Result}", result);


            // Assert
            UnauthorizedObjectResult unauthorizedResult = Assert.IsType<UnauthorizedObjectResult>(result);
            Log.Information("Verified if the response is of type UnauthorizedObjectResult");
            Unauthorized loginResponseReturned = Assert.IsType<Unauthorized>(unauthorizedResult.Value);
            Log.Information("Verified if the response object is of type Unauthorized");
            Assert.Equal(ResponseConstants.PasswordExpired, loginResponseReturned.Detail);
            Log.Information("Verified if the response object's Detail is PasswordExpired");
        }


        [Fact]
        public async Task LoginAsync_ShouldReturn_UnUnauthorizedResponse_WhenPasswordIsIncorrect()
        {
            // Arrange
            LoginRequest loginModel = _fixture.Create<LoginRequest>();
            _mockAuthBl.Setup(x => x.LoginAsync(It.IsAny<LoginRequest>()))
                .ReturnsAsync(new ServiceResponse<LoginServiceResponse>(ResponseConstants.InvalidPassword, HttpStatusCode.Unauthorized));
            Log.Information("Completed Moqing dependencies");


            // Act
            IActionResult result = await _authController.LoginAsync(loginModel);
            Log.Information("Test result: {@Result}", result);

            // Assert
            UnauthorizedObjectResult unauthorizedResult = Assert.IsType<UnauthorizedObjectResult>(result);
            Log.Information("Verified if the response is of type UnauthorizedObjectResult");
            Unauthorized loginResponseReturned = Assert.IsType<Unauthorized>(unauthorizedResult.Value);
            Log.Information("Verified if the response object is of type Unauthorized");
            Assert.Equal(ResponseConstants.InvalidPassword, loginResponseReturned.Detail);
            Log.Information("Verified if the response object's Detail is InvalidPassword");
        }


        [Fact]
        public async Task LoginAsync_ShouldReturnInternalServerError_WhenSqliteExceptionOccers()
        {
            // Arrange
            LoginRequest loginModel = _fixture.Create<LoginRequest>();
            _mockAuthBl.Setup(x => x.LoginAsync(It.IsAny<LoginRequest>()))
                .ThrowsAsync(new SqliteException(nameof(StatusCodes.Status500InternalServerError), StatusCodes.Status500InternalServerError));
            Log.Information("Completed Moqing dependencies");


            // Act & Assert
            await Assert.ThrowsAsync<SqliteException>(() => _authController.LoginAsync(loginModel));
            Log.Information("Verified if the method throwing SqliteException for any db related issues");
        }



        // ValidateSessionAsync
        [Fact]
        public async Task ValidateSessionAsync_ShouldReturnUnauthorized_WhenSessionId_IsNullOrWhiteSpace()
        {
            // Arrange
            List<Claim> claims = new List<Claim>
                {
                    new Claim(ResponseHeaders.SSOSessionId, string.Empty),
                    new Claim(ClaimTypes.Expiration, DateTime.UtcNow.AddMinutes(30).ToString("o"))
                };
            ClaimsIdentity identity = new ClaimsIdentity(claims, CookieAuthenticationDefaults.AuthenticationScheme);
            ClaimsPrincipal claimsPrincipal = new ClaimsPrincipal(identity);
            _mockHttpContext.SetupGet(x => x.User).Returns(claimsPrincipal);
            Log.Information("Completed Moqing dependencies");

            // Act
            IActionResult result = await _authController.ValidateSessionAsync();
            Log.Information("Test result: {@Result}", result);

            // Assert
            UnauthorizedObjectResult unauthorizedResult = Assert.IsType<UnauthorizedObjectResult>(result);
            Log.Information("Verified if the response is of type UnauthorizedObjectResult");
            Unauthorized validateSessionResponseReturned = Assert.IsType<Unauthorized>(unauthorizedResult.Value);
            Log.Information("Verified if the response object is of type Unauthorized");
            Assert.Equal(ResponseConstants.InvalidSessionId, validateSessionResponseReturned.Detail);
            Log.Information("Verified if the response object's Detail is InvalidSessionId");
        }


        [Fact]
        public async Task ValidateSessionAsync_ShouldReturnUnauthorized_WhenSessionIsPassed()
        {
            // Arrange
            _mockHttpContext.SetupGet(x => x.User).Returns((ClaimsPrincipal?)null);
            Log.Information("Completed Moqing dependencies");

            // Act
            IActionResult result = await _authController.ValidateSessionAsync();
            Log.Information("Test result: {@Result}", result);

            // Assert
            UnauthorizedObjectResult unauthorizedResult = Assert.IsType<UnauthorizedObjectResult>(result);
            Log.Information("Verified if the response is of type UnauthorizedObjectResult");
            Unauthorized validateSessionResponseReturned = Assert.IsType<Unauthorized>(unauthorizedResult.Value);
            Log.Information("Verified if the response object is of type Unauthorized");
            Assert.Equal(ResponseConstants.InvalidSessionId, validateSessionResponseReturned.Detail);
            Log.Information("Verified if the response object's Detail is InvalidSessionId");
        }


        [Fact]
        public async Task ValidateSessionAsync_ShouldReturn_OkResponse_WithToken()
        {
            // Arrange
            LoginServiceResponse loginResponse = _fixture
                .Build<LoginServiceResponse>()
                .With(x => x.IsPasswordExpired, false)
                .Create();
            _mockAuthBl.Setup(x => x.ValidateSessionAsync(It.IsAny<string>()))
                .ReturnsAsync(new ServiceResponse<LoginServiceResponse>(ResponseConstants.Success, HttpStatusCode.OK, loginResponse));

            Log.Information("Completed Moqing dependencies");


            // Act
            IActionResult result = await _authController.ValidateSessionAsync();
            Log.Information("Test result: {@Result}", result);

            // Assert
            OkObjectResult okResult = Assert.IsType<OkObjectResult>(result);
            Log.Information("Verified if the response is of type OkObjectResult");
            LoginResponse validateSessionResponseReturned = Assert.IsType<LoginResponse>(okResult.Value);
            Log.Information("Verified if the response object is of type LoginResponse");
            Assert.Equal(loginResponse.Token, validateSessionResponseReturned.Token);
            Log.Information("Verified that mocked token is returned");
        }


        [Fact]
        public async Task ValidateSessionAsync_ShouldReturn_UnauthorizedResponse_If_SessionId_IsInvalid()
        {
            // Arrange
            string sessionId = _fixture.Create<string>();
            _mockAuthBl.Setup(x => x.ValidateSessionAsync(It.IsAny<string>()))
                .ReturnsAsync(new ServiceResponse<LoginServiceResponse>(ResponseConstants.InvalidSessionId, HttpStatusCode.Unauthorized));
            Log.Information("Completed Moqing dependencies");

            // Act
            IActionResult result = await _authController.ValidateSessionAsync();
            Log.Information("Test result: {@Result}", result);

            // Assert
            UnauthorizedObjectResult unauthorizedResult = Assert.IsType<UnauthorizedObjectResult>(result);
            Log.Information("Verified if the response is of type UnauthorizedObjectResult");
            Unauthorized validateSessionResponseReturned = Assert.IsType<Unauthorized>(unauthorizedResult.Value);
            Log.Information("Verified if the response object is of type Unauthorized");
            Assert.Equal(ResponseConstants.InvalidSessionId, validateSessionResponseReturned.Detail);
            Log.Information("Verified if the response object's Detail is InvalidSessionId");
        }


        [Fact]
        public async Task ValidateSessionAsync_ShouldReturn_UnauthorizedResponse_If_SessionId_IsNullOrWhiteSpace()
        {
            // Arrange
            _mockAuthBl.Setup(x => x.ValidateSessionAsync(It.IsAny<string>()))
                .ReturnsAsync(new ServiceResponse<LoginServiceResponse>(ResponseConstants.InvalidSessionId, HttpStatusCode.Unauthorized));
            string sessionId = "";
            List<Claim> claims = new List<Claim>
                {
                    new Claim(ResponseHeaders.SSOSessionId, sessionId),
                    new Claim(ClaimTypes.Expiration, DateTime.UtcNow.AddMinutes(30).ToString("o"))
                };
            ClaimsIdentity identity = new ClaimsIdentity(claims, CookieAuthenticationDefaults.AuthenticationScheme);
            ClaimsPrincipal claimsPrincipal = new ClaimsPrincipal(identity);
            _mockHttpContext.SetupGet(x => x.User).Returns(claimsPrincipal);
            Log.Information("Completed Moqing dependencies");

            // Act
            IActionResult result = await _authController.ValidateSessionAsync();
            Log.Information("Test result: {@Result}", result);

            // Assert
            UnauthorizedObjectResult unauthorizedResult = Assert.IsType<UnauthorizedObjectResult>(result);
            Log.Information("Verified if the response is of type UnauthorizedObjectResult");
            Unauthorized validateSessionResponseReturned = Assert.IsType<Unauthorized>(unauthorizedResult.Value);
            Log.Information("Verified if the response object is of type Unauthorized");
            Assert.Equal(ResponseConstants.InvalidSessionId, validateSessionResponseReturned.Detail);
            Log.Information("Verified if the response object's Detail is InvalidSessionId");
        }


        [Fact]
        public async Task ValidateSessionAsync_ShouldReturn_OkResponse_WithOutSession_IfPasswordIsExpired()
        {
            // Arrange
            string sessionId = _fixture.Create<string>();
            LoginServiceResponse loginResponse = _fixture.Build<LoginServiceResponse>()
                .With(x => x.IsPasswordExpired, true).Create();
            _mockAuthBl.Setup(x => x.ValidateSessionAsync(It.IsAny<string>()))
                .ReturnsAsync(new ServiceResponse<LoginServiceResponse>(ResponseConstants.Success, HttpStatusCode.OK, loginResponse));
            Log.Information("Completed Moqing dependencies");

            // Act
            IActionResult result = await _authController.ValidateSessionAsync();
            Log.Information("Test result: {@Result}", result);

            // Assert
            OkObjectResult okResult = Assert.IsType<OkObjectResult>(result);
            Log.Information("Verified if the response is of type OkObjectResult");
            LoginResponse validateSessionResponseReturned = Assert.IsType<LoginResponse>(okResult.Value);
            Log.Information("Verified if the response object is of type LoginResponse");
            Assert.Equal(loginResponse.Token, validateSessionResponseReturned.Token);
            Log.Information("Verified that mocked token is returned");
            Microsoft.Extensions.Primitives.StringValues setCookieHeader = _httpContext.Response.Headers["Set-Cookie"];
            Assert.DoesNotContain(ResponseHeaders.SSOSessionId, setCookieHeader.ToString());
            Log.Information("Verified if the mocked token is returned from action method invocation");
        }


        [Fact]
        public async Task ValidateSessionAsync_ShouldReturnInternalServerError_WhenSqliteExceptionOccers()
        {
            // Arrange
            string sessionId = _fixture.Create<string>();
            _mockAuthBl.Setup(x => x.ValidateSessionAsync(It.IsAny<string>()))
                .ThrowsAsync(new SqliteException(nameof(StatusCodes.Status500InternalServerError), StatusCodes.Status500InternalServerError));
            Log.Information("Completed Moqing dependencies");


            // Act & Assert
            await Assert.ThrowsAsync<SqliteException>(() => _authController.ValidateSessionAsync());
            Log.Information("Verified if the method throwing SqliteException for any db related issues");
        }



        // Logout API Test Cases
        // Positive Test Scenario
        // Test 1: Valid Session ID with Active Session
        [Fact]
        public async Task Logout_ValidSessionId_ReturnsNoContent()
        {
            // Arrange
            _mockAuthBl.Setup(x => x.ClearUserSessionAsync(It.IsAny<string>()))
                .ReturnsAsync(true);
            Log.Information("Completed Moqing dependencies");

            // Act
            ActionResult result = await _authController.Logout();
            Log.Information("Test result: {@Result}", result);

            // Assert
            Assert.IsType<NoContentResult>(result);
            Log.Information("Verified if the response is of type NoContentResult");
            _mockAuthBl.Verify(x => x.ClearUserSessionAsync(It.IsAny<string>()), Times.Once);
            Log.Information("Verified if the valid session is properly logged out");
        }

        // Negative Test Scenario
        // Test 2: Invalid Session ID (Session not found)
        [Fact]
        public async Task Logout_InvalidSessionId_ReturnsInternalServerError()
        {
            // Arrange
            _mockAuthBl.Setup(x => x.ClearUserSessionAsync(It.IsAny<string>())).ReturnsAsync(false);
            Log.Information("Completed Moqing dependencies");

            // Act
            ActionResult result = await _authController.Logout();
            Log.Information("Test result: {@Result}", result);

            // Assert
            ObjectResult objectResult = Assert.IsType<ObjectResult>(result);
            Log.Information("Verified if the response is of type ObjectResult");
            Assert.Equal((int)HttpStatusCode.InternalServerError, objectResult.StatusCode);
            Log.Information("Verified if the response status is InternalServerError:500");
            _mockAuthBl.Verify(x => x.ClearUserSessionAsync(It.IsAny<string>()), Times.Once);
            Log.Information("Verified if the method returns NotFound for an invalid session");
        }

        // Positive Test Scenario
        // Test 3: Valid Session ID, but No Active Session
        [Fact]
        public async Task Logout_ValidSessionId_NoSession_ReturnsNotFound()
        {
            // Arrange
            _mockHttpContext.SetupGet(x => x.User).Returns((ClaimsPrincipal?)null);
            Log.Information("Completed Moqing dependencies");

            // Act
            ActionResult result = await _authController.Logout();
            Log.Information("Test result: {@Result}", result);

            // Assert
            Assert.IsType<NoContentResult>(result);
            Log.Information("Verified if the response is of type NoContentResult");
            _mockAuthBl.Verify(x => x.ClearUserSessionAsync(It.IsAny<string>()), Times.Never);
            Log.Information("Verified if the method returns NotFound for a valid session with no associated active session");
        }

        // Negative Test Scenario
        // Test 4: Empty Session ID Provided
        [Fact]
        public async Task Logout_EmptyOrNullSessionId_ReturnsBadRequest()
        {
            // Arrange
            List<Claim> claims = new List<Claim>
                {
                    new Claim(ClaimTypes.Expiration, DateTime.UtcNow.AddMinutes(30).ToString("o"))
                };
            ClaimsIdentity identity = new ClaimsIdentity(claims, CookieAuthenticationDefaults.AuthenticationScheme);
            ClaimsPrincipal claimsPrincipal = new ClaimsPrincipal(identity);
            _mockHttpContext.SetupGet(x => x.User).Returns(claimsPrincipal);
            Log.Information("Completed Moqing dependencies");

            // Act
            ActionResult result = await _authController.Logout();
            Log.Information("Test result: {@Result}", result);

            // Assert
            Assert.IsType<BadRequestObjectResult>(result);
            Log.Information("Verified if the response is of type BadRequestObjectResult");
            _mockAuthBl.Verify(x => x.ClearUserSessionAsync(It.IsAny<string>()), Times.Never);
            Log.Information("Verified if the method returns NotFound for an empty session ID");
        }
    }
}
